+++ tdrive/backend/node/src/core/platform/services/database/services/orm/connectors/postgres/postgres-query-builder.ts
@@ -83,8 +83,17 @@ export class PostgresQueryBuilder {
 
       // === IN ===
       options.$in?.forEach(e => {
-        whereClause += `${e[0]} IN ($${e[1].map(() => idx++).join(",$")}) AND `;
-        values.push(...e[1]);
+        const [column, valuesArray] = e;
+
+        // throw an error if the values array is empty
+        if (!valuesArray || valuesArray.length === 0) {
+          throw new Error(
+            `The $in operator for column '${column}' requires a non-empty array of values.`,
+          );
+        }
+
+        whereClause += `${column} IN ($${valuesArray.map(() => idx++).join(",$")}) AND `;
+        values.push(...valuesArray);
       });
 
       // === LIKE ====