+++ tdrive/backend/node/src/services/documents/services/index.ts
@@ -30,6 +30,7 @@ import {
   DriveFileAccessLevel,
   DriveItemDetails,
   DriveTdriveTab,
+  NotificationActionType,
   RootType,
   SearchDocumentsOptions,
   TrashType,
@@ -73,9 +74,9 @@ export class DocumentsService {
   userRepository: Repository<User>;
   ROOT: RootType = "root";
   TRASH: TrashType = "trash";
-  quotaEnabled: boolean = getConfigOrDefault("drive.featureUserQuota", false);
-  defaultQuota: number = getConfigOrDefault("drive.defaultUserQuota", 0);
-  manageAccessEnabled: boolean = getConfigOrDefault("drive.featureManageAccess", false);
+  quotaEnabled: boolean = getConfigOrDefault<boolean>("drive.featureUserQuota", false);
+  defaultQuota: number = getConfigOrDefault<number>("drive.defaultUserQuota", 0);
+  manageAccessEnabled: boolean = getConfigOrDefault<boolean>("drive.featureManageAccess", false);
   logger: TdriveLogger = getLogger("Documents Service");
 
   async init(): Promise<this> {
@@ -457,6 +458,7 @@ export class DocumentsService {
             gr.services.documents.engine.notifyDocumentShared({
               context,
               item: driveItem,
+              type: NotificationActionType.UPDATE,
               notificationEmitter: context.user.id,
               notificationReceiver: parentItem.creator,
             });
@@ -478,7 +480,7 @@ export class DocumentsService {
       await updateItemSize(driveItem.parent_id, this.repository, context);
 
       // If AV feature is enabled, scan the file
-      if (globalResolver.services.av?.avEnabled && version) {
+      if (!driveItem.is_directory && globalResolver.services.av?.avEnabled && version) {
         try {
           driveItem.av_status = await globalResolver.services.av.scanDocument(
             driveItem,
@@ -628,6 +630,7 @@ export class DocumentsService {
                 gr.services.documents.engine.notifyDocumentShared({
                   context,
                   item,
+                  type: NotificationActionType.DIRECT,
                   notificationEmitter: context.user.id,
                   notificationReceiver: sharedWith[0].id,
                 });
@@ -1004,6 +1007,7 @@ export class DocumentsService {
         gr.services.documents.engine.notifyDocumentVersionUpdated({
           context,
           item,
+          type: NotificationActionType.UPDATE,
           notificationEmitter: context.user.id,
           notificationReceiver: item.creator,
         });
@@ -1115,7 +1119,8 @@ export class DocumentsService {
 
       // Check files in the current directory
       const maliciousFiles = entities.filter(
-        child => !child.is_directory && child.av_status !== "safe",
+        child =>
+          !child.is_directory && child.av_status && !["uploaded", "safe"].includes(child.av_status),
       );
 
       if (maliciousFiles.length > 0) {
@@ -1761,6 +1766,7 @@ export class DocumentsService {
     await gr.services.documents.engine.notifyDocumentAVScanAlert({
       context,
       item,
+      type: NotificationActionType.DIRECT,
       notificationEmitter: context.user.id,
       notificationReceiver: context.user.id,
     });