+++ tdrive/frontend/src/app/atoms/button/button.tsx
@@ -22,6 +22,7 @@ interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   disabled?: boolean;
   children?: React.ReactNode;
   shortcut?: string;
+  testClassId?: string;
 }
 
 export const Button = (props: ButtonProps) => {
@@ -86,6 +87,7 @@ export const Button = (props: ButtonProps) => {
       },
     });
   }, [props.onClick, props.shortcut]);
+  const testId = props.testClassId ? `testid:${props.testClassId}` : '';
 
   return (
     <button
@@ -95,10 +97,12 @@ export const Button = (props: ButtonProps) => {
         ' inline-flex items-center px-4 py-2 border font-medium rounded-md focus:outline-none ' +
         className +
         ' ' +
-        props.className
+        props.className +
+        ' ' +
+        testId
       }
       disabled={disabled}
-      {..._.omit(props, 'loading', 'children', 'className', 'icon', 'iconSize')}
+      {..._.omit(props, 'loading', 'children', 'className', 'icon', 'iconSize', 'testClassId')}
     >
       {props.loading && (
         <>