+++ tdrive/frontend/src/app/views/client/body/drive/browser.tsx
@@ -77,9 +77,9 @@ export default memo(
       ? (user?.companies || []).find(company => company?.company.id === companyId)?.role
       : 'member';
     setTdriveTabToken(tdriveTabContextToken || null);
-    const [ filter ] = useRecoilState(SharedWithMeFilterState);
-    const { viewId, dirId } = useRouteState();
-    const [sortLabel] = useRecoilState(DriveItemSort)
+    const [filter] = useRecoilState(SharedWithMeFilterState);
+    const { viewId, dirId, itemId } = useRouteState();
+    const [sortLabel] = useRecoilState(DriveItemSort);
     const [parentId, _setParentId] = useRecoilState(
       DriveCurrentFolderAtom({
         context: context,
@@ -94,8 +94,6 @@ export default memo(
         history.push(RouterServices.generateRouteFromState({ viewId: parentId }));
     }, [viewId, dirId]);
 
-
-
     const [loadingParentChange, setLoadingParentChange] = useState(false);
     const {
       sharedWithMe,
@@ -108,6 +106,7 @@ export default memo(
       loading: loadingParent,
       path,
       loadNextPage,
+      paginateItem,
     } = useDriveItem(parentId);
     const { uploadTree } = useDriveUpload();
 
@@ -145,12 +144,13 @@ export default memo(
       if (item?.id) setUploadModalState({ open: true, parent_id: item.id });
     }, [item?.id, setUploadModalState]);
 
-    const items = item?.is_directory === false
+    const items =
+      item?.is_directory === false
         ? //We use this hack for public shared single file
           item
           ? [item]
           : []
-        : children
+        : children;
 
     const documents = items.filter(i => !i.is_directory);
 
@@ -241,23 +241,73 @@ export default memo(
     }
 
     // Infinite scroll
-    const scrollViwer = useRef<HTMLDivElement>(null);
+    const scrollViewer = useRef<HTMLDivElement>(null);
 
     const handleScroll = async () => {
-      const scrollTop = scrollViwer.current?.scrollTop || 0;
-      const scrollHeight = scrollViwer.current?.scrollHeight || 0;
-      const clientHeight = scrollViwer.current?.clientHeight || 0;
+      const scrollTop = scrollViewer.current?.scrollTop || 0;
+      const scrollHeight = scrollViewer.current?.scrollHeight || 0;
+      const clientHeight = scrollViewer.current?.clientHeight || 0;
       if (scrollTop > 0 && scrollTop + clientHeight >= scrollHeight) {
         await loadNextPage(parentId);
       }
     };
 
     useEffect(() => {
-      if(!loading) scrollViwer.current?.addEventListener('scroll', handleScroll, { passive: true });
+      if (!loading)
+        scrollViewer.current?.addEventListener('scroll', handleScroll, { passive: true });
       return () => {
-        scrollViwer.current?.removeEventListener('scroll', handleScroll);
+        scrollViewer.current?.removeEventListener('scroll', handleScroll);
       };
     }, [parentId, loading]);
+
+    // Scroll to item in view
+    const scrollTillItemInView = itemId && itemId?.length > 0;
+    const scrollItemId = itemId || '';
+
+    useEffect(() => {
+      const itemInChildren = children.find(item => item.id === scrollItemId);
+      if (!loading && scrollTillItemInView && !itemInChildren) {
+        scrollViewer.current?.scrollTo(0, scrollViewer.current?.scrollHeight);
+      } else {
+        if (!loading && itemInChildren) {
+          // scroll to preview item using id for current preview routes
+          const element = document.getElementById(`DR-${scrollItemId}`);
+          element?.scrollIntoView({ behavior: 'smooth', block: 'center' });
+
+          // set it as checked to indicate it is in view
+          setChecked({ [scrollItemId]: true });
+        }
+      }
+    }, [loading, children]);
+
+    // Determine the number of items that can fit within the scroll viewer's visible area before the scrollbar appears.
+    const getItemsPerPage = () => {
+      const scrollViewerElement = scrollViewer?.current || null;
+      const itemHeight = scrollViewerElement?.firstElementChild?.clientHeight || 0;
+      const viewerHeight = scrollViewerElement?.clientHeight || 0;
+      return itemHeight > 0 ? Math.ceil(viewerHeight / itemHeight) : 0;
+    };
+
+    const [itemsPerPage, setItemsPerPage] = useState(0);
+
+    useEffect(() => {
+      const handleResize = () => {
+        setItemsPerPage(getItemsPerPage());
+      };
+      handleResize(); // intially set the items per page for the current view
+      window.addEventListener('resize', handleResize);
+      return () => window.removeEventListener('resize', handleResize);
+    }, [getItemsPerPage]);
+
+    // Load additional pages as needed to ensure the scrollbar remains visible
+    useEffect(() => {
+      const currentPage = Math.floor((paginateItem?.page || 1) / (paginateItem?.limit || 1));
+      const targetPages = Math.ceil(itemsPerPage / (paginateItem?.limit || 1));
+
+      if (!loading && currentPage < targetPages) {
+        loadNextPage(parentId);
+      }
+    }, [paginateItem, loading, parentId, itemsPerPage]);
 
     return (
       <>
@@ -289,6 +339,7 @@ export default memo(
             onDragOver={handleDragOver}
             onDrop={handleDrop}
             disabled={inTrash || access === 'read'}
+            testClassId="browser-upload-zone"
           >
             {role == 'admin' && <UsersModal />}
             <VersionsModal />
@@ -307,7 +358,7 @@ export default memo(
                 (loading && (!items?.length || loadingParentChange) ? 'opacity-50 ' : '')
               }
             >
-              <div className="flex flex-row shrink-0 items-center mb-4">
+              <div className={`flex flex-row shrink-0 items-center mb-4 ${!sharedWithMe ? 'flex-wrap' : ''}`}>
                 {sharedWithMe ? (
                   <div>
                     <Title className="mb-4 block">
@@ -324,8 +375,11 @@ export default memo(
                               buildFileTypeContextMenu(),
                               { x: evt.clientX, y: evt.clientY },
                               'center',
+                              undefined,
+                              "browser-share-with-me-menu-file-type"
                             );
                           }}
+                          testClassId="button-open-menu-file-type"
                         >
                           <span>
                             {filter.mimeType.key && filter.mimeType.key != 'All'
@@ -344,8 +398,11 @@ export default memo(
                               buildPeopleContextMen(),
                               { x: evt.clientX, y: evt.clientY },
                               'center',
+                              undefined,
+                              "browser-share-with-me-menu-people"
                             );
                           }}
+                          testClassId="button-open-menu-people"
                         >
                           <span>{Languages.t('scenes.app.shared_with_me.people')}</span>
                           <ChevronDownIcon className="h-4 w-4 ml-2 -mr-1" />
@@ -361,8 +418,11 @@ export default memo(
                               buildDateContextMenu(),
                               { x: evt.clientX, y: evt.clientY },
                               'center',
+                              undefined,
+                              "browser-share-with-me-menu-last-modified"
                             );
                           }}
+                          testClassId="button-open-menu-last-modified"
                         >
                           <span>
                             {filter.date.key && filter.date.key != 'All'
@@ -390,22 +450,24 @@ export default memo(
                   </BaseSmall>
                 )}
 
-                <Menu menu={() => onBuildSortContextMenu()} sortData={sortLabel}>
+                <Menu menu={() => onBuildSortContextMenu()} sortData={sortLabel} testClassId="browser-menu-sorting">
                   {' '}
-                  <Button theme="outline" className="ml-4 flex flex-row items-center">
+                  <Button theme="outline" className="ml-4 flex flex-row items-center" testClassId="button-sorting">
                     <SortIcon
                       className={`h-4 w-4 mr-2 -ml-1 ${
                         sortLabel.order === 'asc' ? 'transform rotate-180' : ''
                       }`}
                     />
-                    <span>{Languages.t('components.item_context_menu.sorting.selected.' + sortLabel.by)}</span>
+                    <span>
+                      {Languages.t('components.item_context_menu.sorting.selected.' + sortLabel.by)}
+                    </span>
                     <ChevronDownIcon className="h-4 w-4 ml-2 -mr-1" />
                   </Button>
                 </Menu>
                 {viewId !== 'shared_with_me' && (
-                  <Menu menu={() => onBuildContextMenu(details)}>
+                  <Menu menu={() => onBuildContextMenu(details)} testClassId="browser-menu-more">
                     {' '}
-                    <Button theme="secondary" className="ml-4 flex flex-row items-center">
+                    <Button theme="secondary" className="ml-4 flex flex-row items-center" testClassId="button-more">
                       <span>
                         {selectedCount > 1
                           ? `${selectedCount} items`
@@ -419,7 +481,7 @@ export default memo(
               </div>
 
               <DndContext sensors={sensors} onDragEnd={handleDragEnd} onDragStart={handleDragStart}>
-                <div className="grow overflow-auto" ref={scrollViwer}>
+                <div className="grow overflow-auto" ref={scrollViewer}>
                   {items.length === 0 && !loading && (
                     <div className="mt-4 text-center border-2 border-dashed rounded-md p-8">
                       <Subtitle className="block mb-2">
@@ -433,6 +495,7 @@ export default memo(
                             onClick={() => uploadItemModal()}
                             theme="primary"
                             className="mt-4"
+                            testClassId="button-add-doc"
                           >
                             {Languages.t('scenes.app.drive.add_doc')}
                           </Button>