+++ tdrive/frontend/src/app/views/client/body/drive/context-menu.tsx
@@ -67,7 +67,9 @@ export const useOnBuildContextMenu = (
         const isPersonal = item?.scope === 'personal';
         const selectedCount = checked.length;
         const notSafe =
-          !item?.is_directory && !['uploaded', 'safe'].includes(item?.av_status || '');
+          !item?.is_directory &&
+          (item?.av_status || '').length > 0 &&
+          !['uploaded', 'scanning', 'safe'].includes(item?.av_status || '');
 
         let menu: any[] = [];
 
@@ -83,6 +85,7 @@ export const useOnBuildContextMenu = (
             !FeatureTogglesService.isActiveFeatureName(FeatureNames.COMPANY_MANAGE_ACCESS);
           const newMenuActions = [
             {
+              testClassId: 'share',
               type: 'menu',
               icon: 'share-alt',
               text: Languages.t('components.item_context_menu.share'),
@@ -90,6 +93,7 @@ export const useOnBuildContextMenu = (
               onClick: () => setPublicLinkModalState({ open: true, id: item.id }),
             },
             {
+              testClassId: 'manage-access',
               type: 'menu',
               icon: 'users-alt',
               text: Languages.t('components.item_context_menu.manage_access'),
@@ -97,6 +101,7 @@ export const useOnBuildContextMenu = (
               onClick: () => setAccessModalState({ open: true, id: item.id }),
             },
             {
+              testClassId: 'rescan-document',
               type: 'menu',
               icon: 'shield-check',
               text: Languages.t('components.item_context_menu.rescan_document'),
@@ -113,6 +118,7 @@ export const useOnBuildContextMenu = (
                 (notSafe && !(item.av_status === 'scan_failed')),
             },
             {
+              testClassId: 'download',
               type: 'menu',
               icon: 'download-alt',
               text: Languages.t('components.item_context_menu.download'),
@@ -139,6 +145,7 @@ export const useOnBuildContextMenu = (
             }, // */
             { type: 'separator', hide: notSafe },
             {
+              testClassId: 'move',
               type: 'menu',
               icon: 'folder-question',
               text: Languages.t('components.item_context_menu.move'),
@@ -164,6 +171,7 @@ export const useOnBuildContextMenu = (
                 }),
             },
             {
+              testClassId: 'rename',
               type: 'menu',
               icon: 'file-edit-alt',
               text: Languages.t('components.item_context_menu.rename'),
@@ -171,6 +179,7 @@ export const useOnBuildContextMenu = (
               onClick: () => setPropertiesModalState({ open: true, id: item.id, inPublicSharing }),
             },
             {
+              testClassId: 'copy-link',
               type: 'menu',
               icon: 'link',
               text: Languages.t('components.item_context_menu.copy_link'),
@@ -187,6 +196,7 @@ export const useOnBuildContextMenu = (
               },
             },
             {
+              testClassId: 'version',
               type: 'menu',
               icon: 'history',
               text: Languages.t('components.item_context_menu.versions'),
@@ -195,6 +205,7 @@ export const useOnBuildContextMenu = (
             },
             { type: 'separator', hide: access !== 'manage' || inTrash || notSafe },
             {
+              testClassId: 'move-to-trash',
               type: 'menu',
               icon: 'trash',
               text: Languages.t('components.item_context_menu.move_to_trash'),
@@ -203,6 +214,7 @@ export const useOnBuildContextMenu = (
               onClick: () => setConfirmTrashModalState({ open: true, items: [item] }),
             },
             {
+              testClassId: 'restore',
               type: 'menu',
               text: Languages.t('components.item_context_menu.restore'),
               className: 'error',
@@ -213,6 +225,7 @@ export const useOnBuildContextMenu = (
               },
             },
             {
+              testClassId: 'delete',
               type: 'menu',
               text: Languages.t('components.item_context_menu.delete'),
               className: 'error',
@@ -229,6 +242,7 @@ export const useOnBuildContextMenu = (
           // Add selected items related menus
           const newMenuActions: any[] = [
             {
+              testClassId: 'move-multiple',
               type: 'menu',
               text: Languages.t('components.item_context_menu.move_multiple'),
               hide: parent.access === 'read' || inTrash,
@@ -253,6 +267,7 @@ export const useOnBuildContextMenu = (
                 }),
             },
             {
+              testClassId: 'download-multiple',
               type: 'menu',
               text: Languages.t('components.item_context_menu.download_multiple'),
               hide: inTrash,
@@ -270,12 +285,14 @@ export const useOnBuildContextMenu = (
               },
             },
             {
+              testClassId: 'clear-selection',
               type: 'menu',
               text: Languages.t('components.item_context_menu.clear_selection'),
               onClick: () => setChecked({}),
             },
             { type: 'separator', hide: parent.access === 'read' || notSafe },
             {
+              testClassId: 'delete-multiple',
               type: 'menu',
               text: Languages.t('components.item_context_menu.delete_multiple'),
               hide: !inTrash || parent.access !== 'manage',
@@ -288,6 +305,7 @@ export const useOnBuildContextMenu = (
               },
             },
             {
+              testClassId: 'move-to-trash-multiple',
               type: 'menu',
               text: Languages.t('components.item_context_menu.to_trash_multiple'),
               hide: inTrash || parent.access !== 'manage',
@@ -308,6 +326,7 @@ export const useOnBuildContextMenu = (
           const newMenuActions: any[] = inTrash
             ? [
                 {
+                  testClassId: 'empty-trash',
                   type: 'menu',
                   text: Languages.t('components.item_context_menu.trash.empty'),
                   className: 'error',
@@ -322,6 +341,7 @@ export const useOnBuildContextMenu = (
               ]
             : [
                 {
+                  testClassId: 'add-documents',
                   type: 'menu',
                   text: Languages.t('components.item_context_menu.add_documents'),
                   hide: inTrash || parent.access === 'read',
@@ -330,6 +350,7 @@ export const useOnBuildContextMenu = (
                     setUploadModalState({ open: true, parent_id: parent?.item?.id }),
                 },
                 {
+                  testClassId: 'download-folder',
                   type: 'menu',
                   text: Languages.t('components.item_context_menu.download_folder'),
                   hide: inTrash,
@@ -347,6 +368,7 @@ export const useOnBuildContextMenu = (
                   },
                 },
                 {
+                  testClassId: 'copy-link',
                   type: 'menu',
                   text: Languages.t('components.item_context_menu.copy_link'),
                   hide: !hasAnyPublicLinkAccess(item),
@@ -359,6 +381,7 @@ export const useOnBuildContextMenu = (
                 },
                 { type: 'separator', hide: parent.item!.id != 'root' },
                 {
+                  testClassId: 'manage-users',
                   type: 'menu',
                   text: Languages.t('components.item_context_menu.manage_users'),
                   hide: parent.item!.id != 'root',
@@ -417,6 +440,7 @@ export const useOnBuildFileTypeContextMenu = () => {
   return useCallback(() => {
     const menuItems = mimeTypes.map(item => {
       return {
+        testClassId: item.key,
         type: 'menu',
         text: item.key,
         onClick: () => {
@@ -442,6 +466,7 @@ export const useOnBuildPeopleContextMenu = () => {
   return useCallback(() => {
     const menuItems = userList.map(user => {
       return {
+        id: user.first_name,
         type: 'menu',
         text: user.first_name,
         onClick: () => {
@@ -464,6 +489,7 @@ export const useOnBuildDateContextMenu = () => {
   return useCallback(() => {
     const menuItems = [
       {
+        testClassId: 'all-date',
         type: 'menu',
         text: Languages.t('components.item_context_menu.all'),
         onClick: () => {
@@ -480,6 +506,7 @@ export const useOnBuildDateContextMenu = () => {
         },
       },
       {
+        testClassId: 'today',
         type: 'menu',
         text: Languages.t('components.item_context_menu.today'),
         onClick: () => {
@@ -496,6 +523,7 @@ export const useOnBuildDateContextMenu = () => {
         },
       },
       {
+        testClassId: 'last-week',
         type: 'menu',
         text: Languages.t('components.item_context_menu.last_week'),
         onClick: () => {
@@ -512,6 +540,7 @@ export const useOnBuildDateContextMenu = () => {
         },
       },
       {
+        testClassId: 'last-month',
         type: 'menu',
         text: Languages.t('components.item_context_menu.last_month'),
         onClick: () => {
@@ -538,11 +567,13 @@ export const useOnBuildFileContextMenu = () => {
     (item: DriveItem) => {
       const menuItems = [
         {
+          testClassId: 'preview',
           type: 'menu',
           text: Languages.t('components.item_context_menu.preview'),
           onClick: () => preview(item),
         },
         {
+          testClassId: 'download',
           type: 'menu',
           text: Languages.t('components.item_context_menu.download'),
           onClick: () => {
@@ -561,6 +592,7 @@ export const useOnBuildSortContextMenu = () => {
   return useCallback(() => {
     const menuItems = [
       {
+        testClassId: 'sorting-by-date',
         type: 'menu',
         text: Languages.t('components.item_context_menu.sorting.by.date'),
         icon: sortItem.by === 'date' ? 'check' : 'sort-check',
@@ -576,6 +608,7 @@ export const useOnBuildSortContextMenu = () => {
         },
       },
       {
+        testClassId: 'sorting-by-name',
         type: 'menu',
         text: Languages.t('components.item_context_menu.sorting.by.name'),
         icon: sortItem.by === 'name' ? 'check' : 'sort-check',
@@ -590,6 +623,7 @@ export const useOnBuildSortContextMenu = () => {
         },
       },
       {
+        testClassId: 'sorting-by-size',
         type: 'menu',
         text: Languages.t('components.item_context_menu.sorting.by.size'),
         icon: sortItem.by === 'size' ? 'check' : 'sort-check',
@@ -605,6 +639,7 @@ export const useOnBuildSortContextMenu = () => {
       },
       {type:"separator"},
       {
+        testClassId: 'sorting-order-asc',
         type: 'menu',
         text: Languages.t('components.item_context_menu.sorting.order.asc'),
         icon: sortItem.order === 'asc' ? 'check' : 'sort-check',
@@ -619,6 +654,7 @@ export const useOnBuildSortContextMenu = () => {
         },
       },
       {
+        testClassId: 'sorting-order-desc',
         type: 'menu',
         text: Languages.t('components.item_context_menu.sorting.order.desc'),
         icon: sortItem.order === 'desc' ? 'check' : 'sort-check',