+++ tdrive/frontend/src/app/views/client/body/drive/documents/document-row.tsx
@@ -1,6 +1,5 @@
 import {
   DotsHorizontalIcon,
-  ShieldCheckIcon,
   ShieldExclamationIcon,
   BanIcon,
 } from '@heroicons/react/outline';
@@ -52,8 +51,10 @@ export const DocumentRow = ({
         (checked
           ? (notSafe ? 'bg-rose-500' : 'bg-blue-500') + ' bg-opacity-10 hover:bg-opacity-25'
           : 'hover:bg-zinc-500 hover:bg-opacity-10 ') +
-        (className || '')
+        (className || '') + ' ' +
+        'testid:document-row'
       }
+      id={`DR-${item.id}`}
       onMouseEnter={() => setHover(true)}
       onMouseLeave={() => setHover(false)}
       onClick={e => {
@@ -77,24 +78,21 @@ export const DocumentRow = ({
           fallback={<DocumentIcon item={item} />}
         />
       </div>
-      <div className="grow text-ellipsis whitespace-nowrap overflow-hidden">
-        <Base className="flex maxWidth100">{item.name}</Base>
+      <div className="md:grow text-ellipsis whitespace-nowrap">
+        <Base className={`block text-ellipsis whitespace-nowrap overflow-hidden ${hasAnyPublicLinkAccess(item) ? 'w-[38px]' : 'w-[74px]'} md:w-full`}>{item.name}</Base>
       </div>
-      <div className="shrink-0 ml-4">
-        {hasAnyPublicLinkAccess(item) && <PublicIcon className="h-5 w-5 text-blue-500" />}
+      <div className="shrink-0 md:ml-4">
+        {hasAnyPublicLinkAccess(item) && <PublicIcon className="h-5 w-5 text-blue-500 ml-4" />}
       </div>
-      <div className="shrink-0 ml-4 mr-12">
+      <div className="shrink-0 ml-4 md:mr-12">
         <BaseSmall>{formatDateShort(item?.last_version_cache?.date_added)}</BaseSmall>
       </div>
-      <div className="shrink-0 ml-4 text-right lg:w-24 sm:w-20 ">
+      <div className="shrink-0 ml-4 mr-4 md:mr-none text-right lg:w-24 sm:w-20 ">
         <BaseSmall>{formatBytes(item.size)}</BaseSmall>
       </div>
       {FeatureTogglesService.isActiveFeatureName(FeatureNames.COMPANY_AV_ENABLED) && (
         <div className="shrink-0 ml-4 text-right lg:w-24 sm:w-20 ">
           <BaseSmall title={Languages.t(`scenes.app.drive.document_row.av_${item?.av_status}`)}>
-            {item?.av_status === 'scanning' && (
-              <ShieldExclamationIcon className="w-5 text-yellow-400" />
-            )}
             {item?.av_status === 'malicious' && (
               <ShieldExclamationIcon className="w-5 text-rose-400" />
             )}
@@ -103,13 +101,14 @@ export const DocumentRow = ({
           </BaseSmall>
         </div>
       )}
-      <div className="shrink-0 ml-4">
-        <Menu menu={onBuildContextMenu}>
+      <div className="shrink-0 ml-auto md:ml-4">
+        <Menu menu={onBuildContextMenu} testClassId="document-row-menu">
           <Button
             theme={'secondary'}
             size="sm"
             className={'!rounded-full '}
             icon={DotsHorizontalIcon}
+            testClassId="document-row-button-open-menu"
           />
         </Menu>
       </div>