+++ tdrive/frontend/src/app/views/client/body/drive/documents/folder-row.tsx
@@ -28,7 +28,8 @@ export const FolderRow = ({
         (checked
           ? 'bg-blue-500 bg-opacity-10 hover:bg-opacity-25 '
           : 'hover:bg-zinc-500 hover:bg-opacity-10 ') +
-        (className || '')
+        (className || '') + ' ' +
+        'testid:folder-row'
       }
       onMouseEnter={() => setHover(true)}
       onMouseLeave={() => setHover(false)}
@@ -46,8 +47,8 @@ export const FolderRow = ({
           fallback={<FolderIcon className="h-5 w-5 shrink-0 text-blue-500" />}
         />
       </div>
-      <div className="grow text-ellipsis whitespace-nowrap overflow-hidden">
-        <Base className="!font-semibold flex maxWidth100">{item.name}</Base>
+      <div className="grow whitespace-nowrap overflow-hidden">
+        <Base className="!font-semibold text-ellipsis overflow-hidden block max-w-full">{item.name}</Base>
       </div>
       <div className="shrink-0 ml-4">
         {hasAnyPublicLinkAccess(item) && (
@@ -58,12 +59,13 @@ export const FolderRow = ({
         <BaseSmall>{formatBytes(item.size)}</BaseSmall>
       </div>
       <div className="shrink-0 ml-4">
-        <Menu menu={onBuildContextMenu}>
+        <Menu menu={onBuildContextMenu} testClassId="folder-row-menu">
           <Button
             theme={'secondary'}
             size="sm"
             className={'!rounded-full '}
             icon={DotsHorizontalIcon}
+            testClassId="folder-row-button-open-menu"
           />
         </Menu>
       </div>