+++ tdrive/frontend/src/app/views/client/body/drive/header-path.tsx
@@ -74,7 +74,7 @@ export const PathRender = ({
   const pathLength = (pathToRender || []).reduce((acc, curr) => acc + curr.name.length, 0);
 
   return (
-    <nav className="overflow-hidden whitespace-nowrap mr-2 pl-px inline-flex">
+    <nav className="overflow-hidden whitespace-nowrap mr-2 pl-px inline-flex md:max-w-[50%] sm:max-w-[20%] testid:header-path">
       {pathLength < 70 ? (
         (pathToRender || [])?.map((a, i) => (
           <PathItem
@@ -133,20 +133,22 @@ const PathItem = ({
   const { access: trashAccess } = useDriveItem('trash');
   const isInSharedWithMe = viewId === 'shared_with_me';
   return (
-    <div className="flex items-center">
+    <div className={`flex items-center ${!first ? 'overflow-hidden' : ''} testid:header-path-item`}>
       <a
         href="#"
-        className="text-sm font-medium text-gray-700 hover:text-blue-600 dark:text-gray-400 dark:hover:text-white"
+        className={`text-sm font-medium text-gray-700 hover:text-blue-600 dark:text-gray-400 dark:hover:text-white ${!first ? 'overflow-hidden' : ''} testid:item-link`}
         onClick={evt => {
           evt.preventDefault();
 
           const trashMenuItems = [
             {
+              testClassId: 'my-trash',
               type: 'menu',
               text: Languages.t('components.header_path.my_trash'),
               onClick: () => onClick('trash_' + user?.id, ''),
             },
             {
+              testClassId: 'shared-trash',
               type: 'menu',
               text: Languages.t('components.header_path.shared_trash'),
               onClick: () => onClick('trash', ''),
@@ -160,7 +162,13 @@ const PathItem = ({
 
           if (first && user?.id) {
             if (viewId?.includes('trash')) {
-              MenusManager.openMenu(trashMenuItems, { x: evt.clientX, y: evt.clientY }, 'center');
+              MenusManager.openMenu(
+                trashMenuItems,
+                { x: evt.clientX, y: evt.clientY },
+                'center',
+                undefined,
+                'menu-trash'
+              );
             } else {
               if (viewId === 'root') {
                 onClick('root', '');
@@ -173,7 +181,7 @@ const PathItem = ({
           }
         }}
       >
-        <Title noColor={last} className={last ? 'text-blue-500' : ''}>
+        <Title noColor={last} className={!first ? 'text-blue-500 inline-block overflow-hidden text-ellipsis max-w-36' : ''}>
           {(() => {
             const isTrash = viewId?.includes('trash_') || viewId === 'trash';
             const fileName = cutFileName(item?.name) || '';