+++ tdrive/frontend/src/app/views/client/body/drive/modals/upload/index.tsx
@@ -40,7 +40,7 @@ export const UploadModal = ({
     <Modal
       open={state.open}
       onClose={() => setState({ ...state, open: false })}
-      className="!max-w-sm"
+      className="!max-w-sm testid:upload-modal"
     >
       <ModalContent
         title={
@@ -69,18 +69,20 @@ export const UploadModal = ({
             as="div"
             {...(!state.type ? slideXTransitionReverted : slideXTransition)}
           >
-          <div className="-m-2" >
-                  <CreateModalOption
-                      icon={<DocumentDownloadIcon className="w-5 h-5" />}
-                      text={Languages.t('components.create_modal.upload_files')}
-                      onClick={() => selectFromDevice()}
-                  />
-                  <CreateModalOption
-                      icon={<FolderDownloadIcon className="w-5 h-5" />}
-                      text={Languages.t('components.create_modal.upload_folders')}
-                      onClick={() => selectFolderFromDevice()}
-                  />
-                </div>
+            <div className="-m-2" >
+              <CreateModalOption
+                icon={<DocumentDownloadIcon className="w-5 h-5" />}
+                text={Languages.t('components.create_modal.upload_files')}
+                onClick={() => selectFromDevice()}
+                testClassId="upload-file-from-device"
+              />
+              <CreateModalOption
+                icon={<FolderDownloadIcon className="w-5 h-5" />}
+                text={Languages.t('components.create_modal.upload_folders')}
+                onClick={() => selectFolderFromDevice()}
+                testClassId="upload-folder-from-device"
+              />
+            </div>
           </Transition>
         </div>
       </ModalContent>
@@ -88,7 +90,12 @@ export const UploadModal = ({
   );
 };
 
-const CreateModalOption = (props: { icon: ReactNode; text: string; onClick: () => void }) => {
+const CreateModalOption = (props: {
+  icon: ReactNode;
+  text: string;
+  onClick: () => void;
+  testClassId?: string;
+}) => {
   // on press enter
   const handleKeyPress = (e: React.KeyboardEvent<HTMLDivElement>) => {
     if (e.key === 'Enter') {
@@ -99,7 +106,7 @@ const CreateModalOption = (props: { icon: ReactNode; text: string; onClick: () =
   return (
     <div
       onClick={props.onClick}
-      className="flex flex-row p-4 dark:bg-zinc-900 dark:text-white bg-zinc-100 hover:bg-opacity-75 cursor-pointer rounded-md m-2 focus:bg-zinc-800 dark:focus:bg-zinc-800 outline-none focus:border-none"
+      className={`flex flex-row p-4 dark:bg-zinc-900 dark:text-white bg-zinc-100 hover:bg-opacity-75 cursor-pointer rounded-md m-2 focus:bg-zinc-800 dark:focus:bg-zinc-800 outline-none focus:border-none testid:${props.testClassId}`}
       tabIndex={0}
       onKeyUp={handleKeyPress}
     >