+++ tdrive/frontend/src/app/views/client/viewer/drive-preview.tsx
@@ -66,8 +66,11 @@ export const DrivePreview: React.FC<DrivePreviewProps> = ({ items }) => {
     if (items.length < 2)
       // eslint-disable-next-line @typescript-eslint/no-empty-function
       return () => {};
-    addShortcut({ shortcut: 'Right', handler: handleSwitchRight });
-    addShortcut({ shortcut: 'Left', handler: handleSwitchLeft });
+
+    if (!status.loading) {
+      addShortcut({ shortcut: 'Right', handler: handleSwitchRight });
+      addShortcut({ shortcut: 'Left', handler: handleSwitchLeft });
+    }
 
     return () => {
       removeShortcut({ shortcut: 'Right', handler: handleSwitchRight });
@@ -99,12 +102,12 @@ export const DrivePreview: React.FC<DrivePreviewProps> = ({ items }) => {
     <Modal
       open={isOpen}
       closable={false}
-      className="bg-black bg-opacity-50 !sm:max-w-none !w-full !rounded-none !p-0"
+      className="bg-black bg-opacity-50 !sm:max-w-none !w-full !rounded-none !p-0 testid:preview-modal"
       style={{ maxWidth: 'none', margin: 0, left: 0, top: 0, height: '100vh' }}
       positioned={false}
     >
       <XIcon
-        className="z-10 cursor-pointer absolute right-5 top-5 w-20 h-20 text-white hover:text-black rounded-full p-1 bg-gray-500 hover:bg-white bg-opacity-25"
+        className="z-10 cursor-pointer absolute right-5 top-5 w-20 h-20 text-white hover:text-black rounded-full p-1 bg-gray-500 hover:bg-white bg-opacity-25 testid:preview-button-close"
         onClick={() => {
           close();
           // small delay to allow the modal to close
@@ -132,10 +135,10 @@ export const DrivePreview: React.FC<DrivePreviewProps> = ({ items }) => {
         </div>
         <div className="z-10 p-5 bg-black w-full flex text-white">
           <div className="grow overflow-hidden text-ellipsis">
-            <Text.Base noColor className="w-full block text-white whitespace-nowrap">
+            <Text.Base noColor className="w-full block text-white whitespace-nowrap testid:preview-file-name">
               {name}
             </Text.Base>
-            <Text.Info className="whitespace-nowrap">
+            <Text.Info className="whitespace-nowrap testid:preview-file-info">
               {formatDate(
                 +(status.details?.item.added || '') ||
                 status.details?.item.last_version_cache.date_added,
@@ -158,6 +161,7 @@ export const DrivePreview: React.FC<DrivePreviewProps> = ({ items }) => {
                   size="lg"
                   icon={ArrowLeftIcon}
                   onClick={ handleSwitchLeft }
+                  testClassId="drive-preview-button-switch-left"
                 />
                 <Button
                   iconSize="lg"
@@ -166,6 +170,7 @@ export const DrivePreview: React.FC<DrivePreviewProps> = ({ items }) => {
                   size="lg"
                   icon={ArrowRightIcon}
                   onClick={ handleSwitchRight }
+                  testClassId="drive-preview-button-switch-right"
                 />
               </>
             }
@@ -178,6 +183,7 @@ export const DrivePreview: React.FC<DrivePreviewProps> = ({ items }) => {
               onClick= {() => {
                 download && (window.location.href = download);
               }}
+              testClassId="drive-preview-button-download"
             />
           </div>
         </div>